Provisioning: move the read-seam getters into the informer package
The getters are produced by the delta-source selectors and consumed by the
controllers, so co-locate them with the source. RepositoryGetter/ConnectionGetter
and their cached/client implementations move from controller/getters.go into the
per-type informer files (repository.go, connection.go); the selectors now return
the local types and no longer import the controller package. The controllers and
their tests reference informer.RepositoryGetter / informer.ConnectionGetter, so
the dependency is now controller -> informer (one way, no cycle).
